Implementing Member Methods part two. Completed.
