+++ src/main.cpp
@@ -1,77 +1,12 @@
 #include <iostream>
 #include <string>
+#include "include/Account.hpp"
 
-/*
-Implementing member methods in one file inline and outside class
-*/
-
-class Account
-{
-
-private:
-    // attributes
-    std::string m_name;
-    double m_balance;
-
-public:
-    
-    // methods declared inline
-    void set_balance(double bal)
-    {
-        m_balance = bal;
-    }
-
-    double get_balance()
-    {
-        return m_balance;
-    }
-
-    // methods will be declared outside class declaration
-    // prototypes only here
-    void set_name(std::string n);
-    std::string get_name();
-
-    bool deposit(double amount);
-    bool withdraw(double amount);
-     
-};
-
-/* 
-Methods declared outside of class below
-Must include the class name and the scope resolution operator ::
-i.e ClassName::method_name
-Public methods have access to private attributes
-*/
-void Account::set_name(std::string n)
-{
-    m_name = n;
-}
-
-std::string Account::get_name()
-{
-    return m_name;
-}
-
-bool Account::deposit(double amount)
-{
-    // if verify amount
-    m_balance += amount;
-    return true;
-}
-
-bool Account::withdraw(double amount)
-{
-    if (m_balance-amount >= 0){
-        m_balance -= amount;
-        return true;
-    } else {
-        return false;
-    }
-}
 
 int main ()
 {
 
+    // Class Account from Account.hpp
     Account frank_account;
     // Must use public method to set m_name
     frank_account.set_name("Frank's Savings Account");
